+++ src/router/index.js
@@ -30,6 +30,7 @@ const routes = [
     component: OrderStatus
   },
   {
+    // ändra till /profile/:uuid när vi fått back-end uuid att funka
     path: "/profile",
     name: "Profile",
     component: Profile
