+++ src/store/modules/getProfile.js
@@ -1,4 +1,4 @@
-const profile = {
+const getProfile = {
   state: {
     uuid: ""
   },
@@ -7,7 +7,25 @@ const profile = {
   },
   actions: {
     // API: GET uuid goes here
+
+    async getUser(ctx, uuid) {
+      const url = "http://localhost:5000/api/user";
+      fetch(url, {
+        method: "GET",
+        body: uuid,
+        headers: { "Content-Type": "application/json" }
+      })
+        .then(response => response.json())
+        .then(data => {
+          if (data) {
+            console.log(data);
+          }
+        })
+        .catch(error => {
+          console.error("Error:", error);
+        });
+    }
   }
 };
 
-export default profile;
+export default getProfile;
