+++ src/store/modules/postOrder.js
@@ -27,6 +27,37 @@ const postOrder = {
         }
         localStorage.setItem("orderStatus", JSON.stringify(state.orderStatus));
       }, 1000);
+    },
+    // addToCart from MENU-view
+    addToCart(state, product) {
+      const productObj = Object.assign({}, product);
+      // check if product is in cart, if it is in cart, add +1 to quantity and set new total price.
+      if (state.cart.includes(product)) {
+        const index = state.cart.findIndex(obj => obj.id === productObj.id);
+        const price = state.cart[index].price;
+
+        state.cart[index].quantity += 1;
+        state.cart[index].totPrice += price;
+      } else {
+        productObj.quantity = 1;
+        productObj.totPrice = productObj.price;
+
+        state.cart.push(productObj);
+      }
+    },
+    // remove/add from CART-component
+    removeOneProduct(state, product) {
+      const index = state.cart.findIndex(obj => obj.id === product.id);
+
+      if (state.cart[index].quantity === 1) {
+        state.cart[index].splice(index, 1);
+      } else {
+        state.cart[index].quantity -= 1;
+      }
+    },
+    addOneProduct(state, product) {
+      const index = state.cart.findIndex(obj => obj.id === product.id);
+      state.cart[index].quantity += 1;
     }
   },
   actions: {
